fixed rendering issue with user flight list
