+++ components/FlightList.js
@@ -1,8 +1,9 @@
-import React, { useState, useEffect, Component } from "react";
+import React, { Component } from "react";
 import { View, ScrollView } from "react-native";
 import { ListItem } from "react-native-elements";
 
 export default function FlightList(props) {
+  console.log("HELLO", props.flightList);
   return (
     <View style={{ flex: 1, backgroundColor: "#fff" }}>
       <ScrollView>
