+++ package.json
@@ -18,9 +18,8 @@
     "react-native-elements": "^3.3.2",
     "react-native-gesture-handler": "^1.10.3",
     "react-native-reanimated": "^2.1.0",
-    "react-native-safe-area-context": "3.1.9",
-    "react-native-snap-carousel": "^3.9.1",
     "react-native-screens": "^3.1.1",
+    "react-native-snap-carousel": "^3.9.1",
     "react-native-vector-icons": "^8.1.0",
     "react-native-web": "~0.13.12",
     "react-redux": "^7.2.3",
